Create a sign up form component
